+++ packages/Thoth.Json.Core/Decode.fs
@@ -80,6 +80,37 @@ module Decode =
             reason
         | _ -> "Error at: `" + path + "`\n" + reason
 
+    // Low-level API
+    module Advanced =
+
+        /// <summary>
+        /// Runs the decoder against the given JSON value.
+        ///
+        /// If the decoder fails, it reports the error prefixed with the given path.
+        ///
+        /// </summary>
+        /// <example>
+        /// <code lang="fsharp">
+        /// module Decode =
+        ///     let fromRootValue (decoder : Decoder&lt;'T&gt;) =
+        ///         Decode.fromValue "$" decoder
+        /// </code>
+        /// </example>
+        /// <param name="path">Path used to report the error</param>
+        /// <param name="decoder">Decoder to apply</param>
+        /// <param name="value">JSON value to decoder</param>
+        /// <returns>
+        /// Returns <c>Ok</c> if the decoder succeeds, otherwise <c>Error</c> with the error message.
+        /// </returns>
+        let fromValue
+            (helpers: IDecoderHelpers<'JsonValue>)
+            (decoder: Decoder<'T>)
+            =
+            fun value ->
+                match decoder.Decode(helpers, value) with
+                | Ok success -> Ok success
+                | Error error -> Error(errorToString helpers error)
+
     /// <summary>
     /// Decode a JSON string into an F# string.
     /// </summary>
@@ -614,34 +645,6 @@ module Decode =
                 else
                     decoder.Decode(helpers, value) |> Result.map Some
         }
-
-    /// <summary>
-    /// Runs the decoder against the given JSON value.
-    ///
-    /// If the decoder fails, it reports the error prefixed with the given path.
-    ///
-    /// </summary>
-    /// <example>
-    /// <code lang="fsharp">
-    /// module Decode =
-    ///     let fromRootValue (decoder : Decoder&lt;'T&gt;) =
-    ///         Decode.fromValue "$" decoder
-    /// </code>
-    /// </example>
-    /// <param name="path">Path used to report the error</param>
-    /// <param name="decoder">Decoder to apply</param>
-    /// <param name="value">JSON value to decoder</param>
-    /// <returns>
-    /// Returns <c>Ok</c> if the decoder succeeds, otherwise <c>Error</c> with the error message.
-    /// </returns>
-    let fromValue
-        (helpers: IDecoderHelpers<'JsonValue>)
-        (decoder: Decoder<'T>)
-        =
-        fun value ->
-            match decoder.Decode(helpers, value) with
-            | Ok success -> Ok success
-            | Error error -> Error(errorToString helpers error)
 
     //////////////////////
     // Data structure ///