+++ tests/Thoth.Json.Tests.Newtonsoft/Main.fs
@@ -4,20 +4,23 @@ open Expecto
 open Thoth.Json.Tests.Testing
 open Thoth.Json.Core
 open Thoth.Json.Newtonsoft
+open Newtonsoft.Json.Linq
 
 type NewtonsoftEncode() =
     interface IEncode with
         override _.toString spaces json = Encode.toString spaces json
 
 type NewtonsoftDecode() =
-    interface IDecode with
+    interface IDecode<JToken> with
+        override _.fromValue decoder = Decode.fromValue decoder
+
         override _.fromString decoder json = Decode.fromString decoder json
 
         override _.unsafeFromString decoder json =
             Decode.unsafeFromString decoder json
 
 type NewtonsoftTestRunner() =
-    inherit TestRunner<Test, obj>()
+    inherit TestRunner<Test, JToken>()
 
     override _.testList = testList
     override _.testCase = testCase
@@ -28,6 +31,10 @@ type NewtonsoftTestRunner() =
     override _.Encode = NewtonsoftEncode()
 
     override _.Decode = NewtonsoftDecode()
+
+    override _.EncoderHelpers = Encode.helpers
+
+    override _.DecoderHelpers = Decode.helpers
 
 [<EntryPoint>]
 let main args =