+++ packages/Thoth.Json.Core/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Rework encoder API to not need a custom DU ([GH-188](https://github.com/thoth-org/Thoth.Json/pull/188/))
+* Rework object encoder helper to support JSON backends not allowing mutating a JSON object
 
 ### Fixed