+++ packages/Thoth.Json.Core/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Rework encoder API to not need a custom DU ([GH-188](https://github.com/thoth-org/Thoth.Json/pull/188/))
 * Rework object encoder helper to support JSON backends not allowing mutating a JSON object
+* Move `Decode.fromValue` to `Decode.Advanced.fromValue`
 
 ### Fixed