+++ packages/Thoth.Json.Core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Add `resizeArray` support ([GH-182](https://github.com/thoth-org/Thoth.Json/issues/182))
+
 ### Changed
 
 * Rework encoder API to not need a custom DU ([GH-188](https://github.com/thoth-org/Thoth.Json/pull/188/))