+++ packages/Thoth.Json.Core/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.3.0 - 2024-06-15
+
 ### Added
 
 * Add `resizeArray` support ([GH-182](https://github.com/thoth-org/Thoth.Json/issues/182))